Revert the renaming of cartopy examples.
Achieved with ``find  . -name "*.py" -not -name "_*" -print -exec git mv {} . \;``
