+++ lib/cartopy/examples/always_circular_stereo.py
