+++ lib/cartopy/examples/arrows.py
