+++ lib/cartopy/examples/aurora_forecast.py
