+++ lib/cartopy/examples/axes_grid_basic.py
