+++ lib/cartopy/examples/barbs.py
@@ -10,7 +10,7 @@ Plotting barbs.
 import matplotlib.pyplot as plt
 
 import cartopy.crs as ccrs
-from cartopy.examples.vector_data.arrows import sample_data
+from cartopy.examples.arrows import sample_data
 
 
 def main():
