+++ lib/cartopy/examples/eccentric_ellipse.py
