+++ lib/cartopy/examples/effects_of_the_ellipse.py
