+++ lib/cartopy/examples/eyja_volcano.py
