+++ lib/cartopy/examples/favicon.py
