+++ lib/cartopy/examples/feature_creation.py
