+++ lib/cartopy/examples/geostationary.py
