+++ lib/cartopy/examples/global_map.py
