+++ lib/cartopy/examples/hurricane_katrina.py
