+++ lib/cartopy/examples/image_tiles.py
