--- lib/cartopy/examples/lines_and_polygons/README.txt
@@ -1,4 +1,0 @@
-Lines and polygons
-------------------
-
-
