--- lib/cartopy/examples/lines_and_polygons/__init__.py
@@ -1,5 +1,0 @@
-"""
-REMOVE
-------
-
-"""
