+++ lib/cartopy/examples/logo.py
