--- lib/cartopy/examples/miscellanea/README.txt
@@ -1,4 +1,0 @@
-Miscellanea
------------
-
-
