--- lib/cartopy/examples/miscellanea/__init__.py
@@ -1,5 +1,0 @@
-"""
-REMOVE
-------
-
-"""
