+++ lib/cartopy/examples/regridding_arrows.py
