+++ lib/cartopy/examples/rotated_pole.py
