--- lib/cartopy/examples/scalar_data/README.txt
@@ -1,4 +1,0 @@
-Scalar data
------------
-
-
