+++ lib/cartopy/examples/star_shaped_boundary.py
