+++ lib/cartopy/examples/tick_labels.py
