+++ lib/cartopy/examples/tissot.py
