+++ lib/cartopy/examples/tube_stations.py
