+++ lib/cartopy/examples/utm_all_zones.py
