--- lib/cartopy/examples/vector_data/README.txt
@@ -1,4 +1,0 @@
-Vector data
------------
-
-
