--- lib/cartopy/examples/vector_data/__init__.py
@@ -1,5 +1,0 @@
-"""
-REMOVE
-------
-
-"""
