--- lib/cartopy/examples/web_services/README.txt
@@ -1,4 +1,0 @@
-Web services
-------------
-
-
