--- lib/cartopy/examples/web_services/__init__.py
@@ -1,5 +1,0 @@
-"""
-REMOVE
-------
-
-"""
