+++ lib/cartopy/examples/wms.py
