+++ lib/cartopy/examples/wmts.py
