+++ lib/cartopy/examples/wmts_time.py
